remove style.js and add style inside each corresponding file
