+++ web/src/Form.js
@@ -1,19 +1,41 @@
 import React, { Component } from "react";
 import TextboxButton from "./TextboxButton";
 import Spinner from "./Spinner";
-import { FormContainer } from "./Style";
+import glamorous from "glamorous";
+
+const FormContainer = glamorous.div(
+  {
+    border: "1px solid rgba(10, 172, 142, 1)",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    overflow: "hidden",
+    backgroundColor: "white",
+    marginTop: "15px"
+  },
+  props => ({
+    height: `${props.height}px`,
+    width: `${props.width}px`,
+    borderRadius: `${props.radius}px`,
+    backgroundColor: `rgba(10,172,142,${props.alpha})`
+  })
+);
 
 class Form extends Component {
   render() {
-    var { onChange, ...other } = this.props; //pass all props except OnChange
     return (
-      <FormContainer {...other}>
-        <Spinner {...this.props} />
+      <FormContainer
+        height={this.props.height}
+        width={this.props.width}
+        radius={this.props.radius}
+        alpha={this.props.alpha}
+      >
+        <Spinner loading={this.props.loading} />
         <TextboxButton
           onSubmit={this.props.onChange}
           placeholder="username"
           prefix="@"
-          {...other}
+          loading={this.props.loading}
         />
       </FormContainer>
     );
