+++ web/src/Spinner.js
@@ -1,10 +1,54 @@
 import React, { Component } from "react";
-import { SpinnerContainer } from "./Style";
+import glamorous from "glamorous";
+import { css } from "glamor";
+
+const skRotate = css.keyframes({
+  "100%": { transform: `rotate(360deg)` }
+});
+
+const skBounce = css.keyframes({
+  "0%": { transform: `scale(0.0)` },
+  "100%": { transform: `scale(0.0)` },
+  "50%": { transform: `scale(1.0)` }
+});
+
+const SpinnerStyle = glamorous.div(
+  {
+    width: "40px",
+    height: "40px",
+    position: "relative",
+    textAlign: "center",
+    animation: `${skRotate} 2.0s infinite linear`
+  },
+  props => props.loading ? { opacity: props.alpha } : { display: "none" }
+);
+
+const Dot1Style = glamorous.div({
+  width: "60%",
+  height: "60%",
+  display: "inline-block",
+  position: "absolute",
+  top: "0",
+  backgroundColor: "#fafafa",
+  borderRadius: "100%",
+  animation: `${skBounce} 2.0s infinite ease-in-out`
+});
+
+const Dot2Style = glamorous(Dot1Style)({
+  top: "auto",
+  bottom: "0",
+  animationDelay: "-1.0s"
+});
 
 // From https://github.com/tobiasahlin/SpinKit
 class Spinner extends Component {
   render() {
-    return <SpinnerContainer {...this.props} />;
+    return (
+      <SpinnerStyle loading={this.props.loading}>
+        <Dot1Style />
+        <Dot2Style />
+      </SpinnerStyle>
+    );
   }
 }
